Java Interface: - JarCreator: Check if file already exists, and throw exception.   + This happens on Windows systems if uniontypes and records have the same case-insensitive name. - DefinitionsCreator: Fix for when OPENMODELICA environment-variable doesn't contain a trailing directory-separator. - The Java testsuite now generates files in the proper order, but doesn't work on Windows due to case-insensitivity issues.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@4799 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/java_interface/src/org/openmodelica/corba/parser/DefinitionsCreator.java
@@ -22,8 +22,9 @@ public class DefinitionsCreator {
     String filename = String.format("%s/%s/%s%s.java", basepath.getAbsolutePath(), basepackage.replace('.', '/'), (packagename != null ? packagename.replace('.', '/')+"/" : ""), classname);
     File f = new File(filename);
     try {
-      f.mkdirs();
-      f.delete();
+      f.getParentFile().mkdirs();
+      if (f.exists())
+        throw new RuntimeException(f + ": File already exists");
       f.createNewFile();
       FileWriter fw = new FileWriter(f);
       fw.write(st.toString());
